Moving Stress abstraction to the direct class
